+++ processor/src/test/fixtures/bad_input/com/example/injector/includes/ExplicitIncludesOfEnclosedFragmentModel.java
@@ -1,0 +1,19 @@
+package com.example.injector.includes;
+
+import sting.Fragment;
+import sting.Injector;
+
+@Injector( includes = ExplicitIncludesOfEnclosedFragmentModel.MyFragment.class )
+interface ExplicitIncludesOfEnclosedFragmentModel
+{
+  Runnable getRunnable();
+
+  @Fragment
+  interface MyFragment
+  {
+    default Runnable provideRunnable()
+    {
+      return null;
+    }
+  }
+}