+++ processor/src/test/java/sting/processor/StingProcessorTest.java
@@ -106,8 +106,6 @@ public final class StingProcessorTest
         new Object[]{ "com.example.injector.gwt.DisableGwtInjectorModel" },
         new Object[]{ "com.example.injector.gwt.EnableGwtInjectorModel" },
 
-        new Object[]{ "com.example.injector.includes.ExplicitIncludesOfNestedModel" },
-
         new Object[]{ "com.example.injector.outputs.BasicOutputModel" },
         new Object[]{ "com.example.injector.outputs.CollectionContainingMultipleInstancesOutputModel" },
         new Object[]{ "com.example.injector.outputs.CollectionOutputModel" },
@@ -588,6 +586,10 @@ public final class StingProcessorTest
 
         new Object[]{ "com.example.injector.includes.BadTypesInIncludesModel",
                       "@Injector target has an includes parameter containing the value java.util.EventListener that is not a type annotated by either @Fragment or @Injectable" },
+        new Object[]{ "com.example.injector.includes.ExplicitIncludesOfEnclosedFragmentModel",
+                      "@Injector target must not include a @Fragment annotated type that is auto-included as it is enclosed within the injector type" },
+        new Object[]{ "com.example.injector.includes.ExplicitIncludesOfEnclosedInjectableModel",
+                      "@Injector target must not include an @Injectable annotated type that is auto-included as it is enclosed within the injector type" },
         new Object[]{ "com.example.injector.includes.InvalidProvider1IncludesModel",
                       "@Injector target has an 'includes' parameter containing the value com.example.injector.includes.InvalidProvider1IncludesModel.MyComponent that is annotated by @com.example.injector.includes.InvalidProvider1IncludesModel.StingProvider(name=\"[FlatEnclosingName]MF1_[SimpleName]_Provider\") that is annotated by an invalid @StingProvider annotation missing a 'value' parameter of type string." },
         new Object[]{ "com.example.injector.includes.InvalidProvider2IncludesModel",