+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Generate an error if an `@Injector` annotated type encloses a type that is not annotated with either `@Injectable` or `@Fragment`.
 * Eliminate a crash when a primitive type dependency was missing and the processor attempted to auto-discover the type.
+* Generate an error if an auto-included type is explicitly included in the includes of an `@Injector`. This means that any `@Injectable` or `@Fragment` directly enclosed within an injector type will generate an error.
 
 ### [v0.09](https://github.com/sting-ioc/sting/tree/v0.09) (2020-03-19) · [Full Changelog](https://github.com/sting-ioc/sting/compare/v0.08...v0.09) · [API Differences](https://sting-ioc.github.io/api-diff?key=sting&old=0.08&new=0.09)